# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Bank Cards API
  description: |-
    REST API системы управления банковскими картами.

        Функциональность:
        - регистрация и аутентификация пользователей
        - управление банковскими картами
        - переводы между картами
        - заявки на активацию и блокировку карт
        - разграничение доступа по ролям (ADMIN / USER)

        Аутентификация осуществляется с помощью JWT (Bearer Token).
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- bearerAuth: []
...
paths:
...
  /api/transactions/transfer:
    post:
      tags:
      - Transactions
      summary: Перевод средств
      description: Перевод средств между картами пользователя
      operationId: transfer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionCreateDto"
        required: true
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Нет прав USER
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "201":
          description: Транзакция выполнена
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TransactionDto"
      security:
      - bearerAuth: []
  /api/cards:
    get:
      tags:
      - Cards
      summary: Получить свои карты (только USER)
      description: Список карт текущего пользователя с фильтрацией
      operationId: getMyCards
      parameters:
      - name: status
        in: query
        description: Статус карты
        required: false
        schema:
          type: string
          enum:
          - NOT_ACTIVE
          - ACTIVE
          - BLOCKED
          - EXPIRED
      - name: expirationBefore
        in: query
        description: Дата окончания до
        required: false
        schema:
          type: string
          format: date
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Нет прав USER
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: Список карт
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageCardDto"
      security:
      - bearerAuth: []
    post:
      tags:
      - Cards
      summary: Создать карту (только ADMIN)
      description: Создание новой карты (только ADMIN)
      operationId: createCard
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardCreateDto"
        required: true
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Нет прав ADMIN
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "201":
          description: Карта создана
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardDto"
      security:
      - bearerAuth: []
...
  /api/transactions:
    get:
      tags:
      - Transactions
      summary: Получить свои транзакции
      description: История транзакций с фильтрацией и пагинацией
      operationId: getMyTransactions
      parameters:
      - name: status
        in: query
        description: Статус транзакции
        required: false
        schema:
          type: string
          enum:
          - SUCCESS
          - FAILED
      - name: from
        in: query
        description: Дата от
        required: false
        schema:
          type: string
          format: date-time
      - name: to
        in: query
        description: Дата до
        required: false
        schema:
          type: string
          format: date-time
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Нет прав USER
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: Список транзакций
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageTransactionDto"
      security:
      - bearerAuth: []
  /api/transactions/admin:
    get:
      tags:
      - Transactions
      summary: Просмотр всех транзакций (ADMIN)
      description: |
        Позволяет администратору получить список всех транзакций в системе.
        Поддерживает фильтрацию по статусу и диапазону дат,
        а также постраничный вывод результатов.
      operationId: getAllTransactions
      parameters:
      - name: status
        in: query
        description: Статус транзакции (необязательный фильтр)
        required: false
        schema:
          type: string
          enum:
          - SUCCESS
          - FAILED
        example: SUCCESS
      - name: from
        in: query
        description: Начальная дата и время (ISO-8601)
        required: false
        schema:
          type: string
          format: date-time
        example: 2026-01-01T00:00:00
      - name: to
        in: query
        description: Конечная дата и время (ISO-8601)
        required: false
        schema:
          type: string
          format: date-time
        example: 2026-01-31T23:59:59
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          type: integer
          default: 20
          minimum: 1
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          items:
            type: string
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Доступ запрещён. Требуется роль ADMIN
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: Список транзакций успешно получен
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageTransactionDto"
      security:
      - bearerAuth: []
  /api/cards/{cardId}:
    get:
      tags:
      - Cards
      summary: Получить свою карту по ID(только USER)
      operationId: getMyCard
      parameters:
      - name: cardId
        in: path
        description: ID карты
        required: true
        schema:
          type: integer
          format: int64
        example: 10
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardDto"
      security:
      - bearerAuth: []
    delete:
      tags:
      - Cards
      summary: Удалить карту(только ADMIN)
      operationId: delete
      parameters:
      - name: cardId
        in: path
        description: ID карты
        required: true
        schema:
          type: integer
          format: int64
        example: 10
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "204":
          description: No Content
      security:
      - bearerAuth: []
  /api/cards/admin:
    get:
      tags:
      - Cards
      summary: Получить все карты (только ADMIN)
      description: Просмотр всех карт в системе с фильтрацией
      operationId: getAllCards
      parameters:
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - NOT_ACTIVE
          - ACTIVE
          - BLOCKED
          - EXPIRED
      - name: expirationBefore
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Нет прав ADMIN
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: Список карт
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageCardDto"
      security:
      - bearerAuth: []
  /api/cards/admin/{cardId}:
    get:
      tags:
      - Cards
      summary: Получить карту по ID (только ADMIN)
      operationId: getCardById
      parameters:
      - name: cardId
        in: path
        description: ID карты
        required: true
        schema:
          type: integer
          format: int64
        example: 10
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardDto"
      security:
      - bearerAuth: []
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        timestamp:
          type: string
          format: date-time
...
    TransactionCreateDto:
      type: object
      description: Данные для перевода между картами
      properties:
        fromCardNumber:
          type: string
          format: uuid
          description: Публичный номер карты отправителя
        toCardNumber:
          type: string
          format: uuid
          description: Публичный номер карты получателя
        amount:
          type: number
          description: Сумма перевода
      required:
      - fromCardNumber
      - toCardNumber
    TransactionDto:
      type: object
      description: Информация о транзакции
      properties:
        id:
          type: integer
          format: int64
          description: ID транзакции
        fromCard:
          type: string
          description: Карта отправителя
        toCard:
          type: string
          description: Карта получателя
        amount:
          type: number
          description: Сумма транзакции
        status:
          type: string
          description: Статус транзакции
          enum:
          - SUCCESS
          - FAILED
        createdAt:
          type: string
          format: date-time
          description: Дата и время создания транзакции
    CardCreateDto:
      type: object
      description: Данные для создания банковской карты
      properties:
        ownerId:
          type: integer
          format: int64
          description: ID владельца карты
          example: 42
        initialBalance:
          type: number
          description: Начальный баланс карты
          example: 1000.0
      required:
      - initialBalance
      - ownerId
    CardDto:
      type: object
      description: Информация о банковской карте
      properties:
        id:
          type: integer
          format: int64
          description: ID карты
          example: 10
        publicNumber:
          type: string
          format: uuid
          description: Публичный номер карты
        ownerId:
          type: integer
          format: int64
          description: ID владельца
          example: 1
        cardNumber:
          type: string
          description: Номер карты
          example: '**** **** **** 1234'
        expirationDate:
          type: string
          format: date
          description: Дата окончания действия карты
          example: 2028-12-31
        status:
          type: string
          description: Статус карты
          enum:
          - NOT_ACTIVE
          - ACTIVE
          - BLOCKED
          - EXPIRED
          example: ACTIVE
        balance:
          type: number
          description: Текущий баланс карты
          example: 750.5
...
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
...
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        paged:
          type: boolean
        sort:
          $ref: "#/components/schemas/SortObject"
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
    PageTransactionDto:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/TransactionDto"
        number:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        pageable:
          $ref: "#/components/schemas/PageableObject"
        empty:
          type: boolean
    PageCardDto:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/CardDto"
        number:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        pageable:
          $ref: "#/components/schemas/PageableObject"
        empty:
          type: boolean
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT